requirements.txt
dnspython==2.2.1
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.28.1
tqdm==4.64.1
urllib3==1.26.12